+++ .gitignore
@@ -3,5 +3,5 @@ __pycache__
 python-virtual-environments
 venv
 statscode
+temp/*
 keys.txt
-tests/temp/*
